requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic<3

langchain==0.3.27
langchain-core==0.3.75
langchain-community==0.3.28
langchain-text-splitters==0.3.9
langchain-google-genai==2.1.10
google-genai>=0.3.0

faiss-cpu==1.12.0
pypdf==6.0.0
numpy<2
python-dotenv==1.1.1
httpx>=0.27
